# project-zer-supply-v2/site/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Project Zer Supply V2",
    "description": "Supply V2 is a dropship procurement platform that splits customer orders into vendor purchase orders, tracks shipments and receipts, resolves disputes, performs invoice matching, and protects every route with authentication, RBAC, and policy decisions.",
    "version": "2.0.0",
    "contact": {
      "name": "Project Zer Platform Team"
    }
  },
  "paths": {
...
    "/purchase-orders/{po_id}/shipments": {
      "post": {
        "tags": [
          "fulfilment"
        ],
        "summary": "Create Shipment",
        "operationId": "create_shipment_purchase_orders__po_id__shipments_post",
        "parameters": [
          {
            "name": "po_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Po Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "tenant_demo",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "user_demo",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "admin",
              "title": "X-Role"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShipmentCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/disputes/{dispute_id}": {
      "get": {
        "tags": [
          "disputes"
        ],
        "summary": "Get Dispute",
        "operationId": "get_dispute_disputes__dispute_id__get",
        "parameters": [
          {
            "name": "dispute_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dispute Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "tenant_demo",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "user_demo",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "admin",
              "title": "X-Role"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/disputes/{dispute_id}/resolve": {
      "post": {
        "tags": [
          "disputes"
        ],
        "summary": "Resolve Dispute",
        "operationId": "resolve_dispute_disputes__dispute_id__resolve_post",
        "parameters": [
          {
            "name": "dispute_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dispute Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "tenant_demo",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "user_demo",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "admin",
              "title": "X-Role"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeResolveIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/purchase-orders/{po_id}/invoices": {
      "post": {
        "tags": [
          "invoices"
        ],
        "summary": "Create Invoice",
        "operationId": "create_invoice_purchase_orders__po_id__invoices_post",
        "parameters": [
          {
            "name": "po_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Po Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "tenant_demo",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "user_demo",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "admin",
              "title": "X-Role"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/purchase-orders/{po_id}/slas": {
      "get": {
        "tags": [
          "invoices"
        ],
        "summary": "List Po Slas",
        "operationId": "list_po_slas_purchase_orders__po_id__slas_get",
        "parameters": [
          {
            "name": "po_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Po Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "tenant_demo",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "user_demo",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "admin",
              "title": "X-Role"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DisputeResolveIn": {
        "properties": {
          "resolution": {
            "type": "string",
            "title": "Resolution"
          }
        },
        "type": "object",
        "required": [
          "resolution"
        ],
        "title": "DisputeResolveIn"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InvoiceCreate": {
        "properties": {
          "invoice_number": {
            "type": "string",
            "title": "Invoice Number"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/InvoiceLineIn"
            },
            "type": "array",
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "invoice_number",
          "lines"
        ],
        "title": "InvoiceCreate"
      },
      "InvoiceLineIn": {
        "properties": {
          "po_line_id": {
            "type": "string",
            "title": "Po Line Id"
          },
          "billed_quantity": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Billed Quantity"
          },
          "billed_unit_price_minor": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Billed Unit Price Minor"
          }
        },
        "type": "object",
        "required": [
          "po_line_id",
          "billed_quantity",
          "billed_unit_price_minor"
        ],
        "title": "InvoiceLineIn"
      },
...
      "ShipmentCreate": {
        "properties": {
          "tracking_number": {
            "type": "string",
            "title": "Tracking Number"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/ShipmentLineIn"
            },
            "type": "array",
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "tracking_number",
          "lines"
        ],
        "title": "ShipmentCreate"
      },
      "ShipmentLineIn": {
        "properties": {
          "po_line_id": {
            "type": "string",
            "title": "Po Line Id"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Quantity"
          }
        },
        "type": "object",
        "required": [
          "po_line_id",
          "quantity"
        ],
        "title": "ShipmentLineIn"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "TenantHeaderAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tenant-Id"
      },
      "InternalKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Internal-Api-Key"
      }
    }
  },
...
}